Created the files for the start of the template
• Changed the name of the public folder to assets – makes a bit more
sense.
• Added some more projects to the content folder to make it easier for
layout purposes.
• Added 'head', 'footer', and 'css' partials to more easily control all
the different template pages.
• Changed the names of the category listing partials to be more
discearnable. So 'categories-list' lists all of the project items from
each category for the homepage, and 'category-list' lists all the
projects for the individual categories.
• Changed the listing of projects on the homepage to use the <article>
tag instead of an <li>.
• Changed the markup of the homepage to be cleaner and use HTML 5 items.
• Removed most of the default stacey css.
• Added the LESS files to control design better, seperated the colors
into their own file and imported it into the screen.less file to be
compiled to screen.css.
 • Added modernizr.js to fix some browser issues with that new fangled
code.