Playing with the fluid layout widths of the home page.
Tried using isotope and masonry to create the effect that I wanted but they wouldn't work due to the fact that both of them did not play nicely with fluid widths for the floating objects. So instead created a set of media queries to resize the portfolio objects on the homepage.