Finished the layout of most of the page.
Set up most of the home page. There is a weird gap on the right side of the thumbnails so I'm gonna have to play with that.

Added an overall @themecolor to the color.less file to make changing the theme color much easier.

Added a favorite things randomizer that pulls from _favoritethings.txt file. This way if you don't know how to code its very easy to go in and add or remove fun facts about yourself, cause anyone knows how to edit a text file.

With this, added the ability to in a .txt file turn on or off the favorite things. just by typing the word yes or no. It takes and applies the text to the class of the object in the template triggering the correct css.

Started on the interior pages.