Some more layout and design updates
Tried fixing the wierd gap on the right side of the thumbnails on the
homepage. Think that I have fixed it by adding a clearing element at
the end of the thumbnails and giving the whole section a white
background.

fixed the interior template pages that had the wrong link back to the
homepage. I copied over all the code from the homepage and removed what
didn't need to show up which caused the problem. Its fixed now.

Also updated the color scheme.